Safari-609.1.11.1/LayoutTests/js/modify-non-references-expected.txt

PASS function f() { g()++; } f.toString() is 'function f() { g()++; }'
PASS function f() { g()--; } f.toString() is 'function f() { g()--; }'
PASS function f() { ++g(); } f.toString() is 'function f() { ++g(); }'
PASS function f() { --g(); } f.toString() is 'function f() { --g(); }'
PASS function f() { g() = 1; } f.toString() is 'function f() { g() = 1; }'
PASS function f() { g() += 1; } f.toString() is 'function f() { g() += 1; }'
PASS g()++ threw exception ReferenceError: Postfix ++ operator applied to value that is not a reference..
PASS g()-- threw exception ReferenceError: Postfix -- operator applied to value that is not a reference..
PASS ++g() threw exception ReferenceError: Prefix ++ operator applied to value that is not a reference..
PASS --g() threw exception ReferenceError: Prefix -- operator applied to value that is not a reference..
PASS g() = 1 threw exception ReferenceError: Left side of assignment is not a reference..
PASS g() += 1 threw exception ReferenceError: Left side of assignment is not a reference..
PASS successfullyParsed is true

TEST COMPLETE
